usermail/src/main/java/com/syswin/temail/usermail/application/UsermailSessionService.java
/*
 * MIT License
 *
 * Copyright (c) 2019 Syswin
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package com.syswin.temail.usermail.application;

import org.apache.commons.codec.digest.DigestUtils;
import org.springframework.stereotype.Component;

@Component
public class UsermailSessionService {

  /**
   * 获取会话id
   *
   * @param from 发件人
   * @param to 收件人
   * @return 会话id
   */
  public String getSessionID(String from, String to) {
    if (null == from || "".equals(from)) {
      throw new IllegalArgumentException("param [from] is illegal");
    }
    if (null == to || "".equals(to)) {
      throw new IllegalArgumentException("param [to] is illegal");
    }
    String value;
    if (from.compareTo(to) > 0) {
      value = from + to;
    } else {
      value = to + from;
    }
    return DigestUtils.md5Hex(value);
  }

}
